// src/shared/api/generated/schemas/projectsSingleProjectResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Smetchik Pro API
 * HTTP API сервиса smethick-pro.
 * OpenAPI spec version: 1.0
 */
import type { ProjectsProjectResponse } from './projectsProjectResponse';

export interface ProjectsSingleProjectResponse {
  project?: ProjectsProjectResponse;
}